requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-bootstrap3==9.0.0
django-decouple==2.1
django-floppyforms==1.7.0
django-material==1.0.0
django-materialize-forms==0.2.0
django-materializecss-form==1.0.1
django-superform==0.3.1
django-tag-parser==3.1
django-toolbelt==0.0.1
django-viewform==0.1.0
django-widget-tweaks==1.4.1
gunicorn==19.7.1
material==0.1
psycopg2==2.7.2
python-decouple==3.0
pytz==2017.2
static3==0.7.0
whitenoise==3.3.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
crispy-forms-materialize==0.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-bootstrap-form==3.3
django-bootstrap3==9.0.0
django-crispy-forms==1.6.1
django-decouple==2.1
django-enumfields==0.9.0
django-filter==1.0.4
django-floppyforms==1.7.0
django-material==1.0.0
django-materialize-form==0.1.1
django-materialize-forms==0.2.0
django-materializecss-form==1.0.1
django-rest-swagger==2.1.2
django-superform==0.3.1
django-tag-parser==3.1
django-toolbelt==0.0.1
django-viewform==0.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
material==0.1
openapi-codec==1.3.2
psycopg2==2.7.2
python-decouple==3.0
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
static3==0.7.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.0
